refactor(apig/channel): deprecate channel resource
