+++ huaweicloud/services/acceptance/deprecated/resource_huaweicloud_apig_vpc_channel_test.go
@@ -1,4 +1,4 @@
-package apig
+package deprecated
 
 import (
 	"fmt"
@@ -7,7 +7,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 
-	"github.com/chnsz/golangsdk/openstack/apigw/dedicated/v2/channels"
+	"github.com/chnsz/golangsdk/openstack/apigw/deprecated/dedicated/v2/channels"
+
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/config"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance"
 	"github.com/huaweicloud/terraform-provider-huaweicloud/huaweicloud/services/acceptance/common"
